Best school projects/Edge Computing/AI mantrap gate security/src/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

fastapi==0.115.11
filterpy==1.4.5
gpiozero==2.0.1
imutils==0.5.4
librosa==0.10.2.post1
mediapipe==0.10.9
numba==0.61.0
numpy==1.26.4
opencv-python==4.11.0.86
paho-mqtt==2.1.0
PyAudio==0.2.14
RPi.GPIO==0.7.1
RPLCD==1.3.1
scipy==1.15.2
smbus2==0.5.0
SpeechRecognition==3.14.1
tensorflow==2.18.0
tflite-runtime==2.14.0
uvicorn==0.34.0
noisereduce
sympy
